finalize seed use in world factory

+++ core/src/main/se/tevej/game/model/factories/BuildingFactory.java
@@ -13,7 +13,7 @@ import main.se.tevej.game.model.resource.Resource;
 import main.se.tevej.game.model.resource.ResourceType;
 
 public class BuildingFactory {
-    public static Entity createBuilding(BuildingType type, float x, float y) throws NoSuchBuildingException {
+    public static Entity createBuilding(BuildingType type, int x, int y) throws NoSuchBuildingException {
         Entity building = new Entity();
         building.add(new PositionComponent(x,y));
         building.add(new SizeComponent(1,1));
